+++ lib/panda_doc/objects/status.rb
@@ -1,0 +1,9 @@
+module PandaDoc
+  module Objects
+    class Status < Virtus::Attribute
+      def coerce(value)
+        value.to_s.split(".").last
+      end
+    end
+  end
+end